move postinstall
