+++ package.json
@@ -80,6 +80,7 @@
     "yui-compressor": "^0.1.3"
   },
   "scripts": {
+    "setup": "node tools/setup.js",
     "postinstall": "node tools/postinstall.js",
     "start": "node app.js",
     "test:browser:integration": "mocha browser/test/integration/issue195.js",
